SYS-2825 Move existing tests
